backend/requirements.txt
django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.26.0
sqlparse==0.3.1
